Current, tested on F4

+++ Core/Inc/Effect.h
@@ -6,6 +6,7 @@
 
 COLOR_GRB GetColorFromPalleteSmooth (uint8_t Angle, PALLETE_ARRAY *PalleteArray);
 COLOR_GRB GetColorFromPalleteSolid (uint8_t Angle, PALLETE_ARRAY *PalleteArray);
-uint8_t LerpHSV (uint8_t a, uint8_t b, uint8_t t);
+uint8_t LerpHSV (uint8_t a, uint8_t b, float t);
+
 
 #endif // _EFFECT_
